+++ src/features/dev-page/dev-page.tsx
@@ -1,0 +1,13 @@
+import React, { useEffect } from "react"
+import styled from "styled-components/macro"
+
+const Wrapper = styled.div`
+  width: 100%;
+  min-height: 20rem;
+  display: flex;
+`
+export const DevPage = () => {
+  useEffect(() => {}, [])
+
+  return <Wrapper></Wrapper>
+}