+++ src/features/tips/tips.styles.ts
@@ -1,9 +1,10 @@
-import styled from "styled-components"
+import styled from "styled-components/macro"
 
 export const Wrapper = styled.span`
   font-size: 2rem;
   color: var(--c-extra-gray);
   margin-bottom: 0;
+  cursor: default;
   margin-top: 3rem;
 `