+++ src/features/user-stats/user-stats.styles.ts
@@ -1,0 +1,14 @@
+import styled from "styled-components/macro"
+
+export const StatsWrapper = styled.div`
+  height: 100%;
+  width: 100%;
+`
+
+export const ChartWrapper = styled.div`
+  height: 80%;
+  width: 100%;
+`
+export const StyledSpan = styled.p`
+  color: var(--c-extra-gray);
+`